requirements.txt
Django==2.2.4
djangorestframework==3.10.2
Pillow==5.2.0
virtualenv==15.1.0
django-filter==2.2.0
gunicorn==19.9.0
whitenoise==4.1.4
model-mommy==2.0.0
ipdb==0.12.2